Move non-generator related stuff to separate files
