+++ src/ggrc/app.py
@@ -102,7 +102,7 @@ def init_permissions_provider():
 
 def init_extra_listeners():
   from ggrc.automapper import register_automapping_listeners
-  from ggrc.snapshoter import register_snapshot_listeners
+  from ggrc.snapshoter.listeners import register_snapshot_listeners
   register_automapping_listeners()
   register_snapshot_listeners()
